requirements.txt
mypy==0.770
mypy-extensions==0.4.3
psycopg2-binary==2.8.4
typed-ast==1.4.1
typing-extensions==3.7.4.1
